requirements.txt
androguard==3.3.5
apturl==0.5.2
arch==5.1.0
asn1crypto==1.4.0
async-generator==1.10
async-timeout==4.0.2
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
boto==2.49.0
Brlapi==0.7.0
catfish==1.4.13
ccsm==0.9.14.1
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
compizconfig-python==0.9.14.1
constantly==15.1.0
cryptography==40.0.2
cssselect==1.1.0
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
decorator==5.1.0
defer==1.0.6
Deprecated==1.2.13
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.6.0
fonttools==4.28.4
future==0.18.2
gpg===1.13.1-unknown
h11==0.13.0
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
imageio==2.19.2
incremental==16.10.1
ipython==7.30.1
ipython-genutils==0.2.0
itchat==1.3.10
itemadapter==0.5.0
itemloaders==1.0.4
jedi==0.18.1
jieba==0.42.1
jmespath==1.0.0
joblib==1.1.0
keras==2.7.0
keyring==18.0.1
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lightdm-gtk-greeter-settings==1.2.2
linecache2==1.0.0
lockfile==0.12.2
louis==3.12.0
lxml==4.7.1
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
menulibre==2.2.1
monotonic==1.5
mugshot==0.4.2
mysqlclient==1.4.4
netifaces==0.10.4
networkx==2.6.3
numpy==1.21.4
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
outcome==1.1.0
packaging==21.3
pandas==1.4.0
paramiko==2.6.0
parsel==1.5.2
parso==0.8.3
patsy==0.5.2
pbr==5.4.5
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
prompt-toolkit==3.0.24
property-cached==1.6.4
Protego==0.2.1
protobuf==3.6.1
psutil==5.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pyaxmlparser==0.3.27
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycryptodomex==3.6.1
pycups==1.9.73
PyDispatcher==2.0.5
pydot==1.4.2
Pygments==2.10.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==23.1.1
pyparsing==3.0.6
pypng==0.0.21
PyQRCode==1.2.1
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2021.3
pyxdg==0.26
PyYAML==5.3.1
queuelib==1.5.0
redis==4.2.2
reportlab==3.5.34
requests==2.22.0
requests-file==1.5.1
requests-unixsocket==0.2.0
scikit-learn==1.1.1
scipy==1.8.1
Scrapy==2.6.1
screen-resolution-extra==0.0.0
seaborn==0.11.2
SecretStorage==2.3.1
selenium==4.1.3
service-identity==18.1.0
sgt-launcher==0.2.5
simplejson==3.16.0
six==1.14.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==1.9.5
ssh-import-id==5.10
statsmodels==0.13.1
system-service==0.3
systemd-python==234
testresources==2.0.0
threadpoolctl==3.1.0
tldextract==3.2.1
traceback2==1.4.0
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
Twisted==18.9.0
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==1.26.9
usb-creator==0.3.7
virtualenv==20.14.1
w3lib==1.21.0
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
wordcloud==1.8.1
wrapt==1.14.0
wsproto==1.1.0
xcffib==0.8.1
xkit==0.0.0
zope.interface==4.7.1
